Feat(global): Initialize project with basic setup
Started out by deleting the existing project to start from scratch. I lost
track of I had actually accomplished in the project, but I knew it wasn't
much. Starting fresh lets me apply what I've learned without having to
backtrack.

I initialized a new Vue project, a new git repo, and a new GitHub repo. I
then changed some of the Vue template. For instance, I switched out the
favicon for an image of Bill Seidman (an inside joke among my friends).
I changed the title of the page to "Mystic V Hold 'em" and deleted the
default HelloWorld component and its references.

Finally, I added an "Usher" component to work on tomorrow. This is going
to be the first thing a user sees when they go to the website. It'll be
responsible for getting the user's name and their initial buy-in.