Feat(SeatUsher): add input for user's name
Started out today by cloning the Github repo to a new machine. It's
embarrassing to admit, but I had to check the Github docs for a reminder
on how to do this. Once that was figured out, though, I went ahead and set
up my environment, i.e. I installed the necessary extensions in VSCode.

I renamed "Usher" to "SeatUsher" per Vue's multi-word-component-name rule
and then hoisted that component into the main App. After that, I started
experimenting with a username prop to collect and store the user's input.

At first, I was struggling to do this with html inputs in the template
section and the props in the script section. I therefore tried adding
a v-model property to the text input instead of a type property. Still
struggled because of a no-mutating-props rule, so had to visit
documentation again. Using Vue's playground example, I saw that the prop
has to be included as a key within the returned object in the data method
of the script section's export. Added that to my code, and sure enough,
it worked.

Next step is to figure out if v-model can pass that info along to be
stored in some sort of array or if I'll have to use something else.