Tweaked SeatUsher component
Trying to change the SeatUsher component so the user has to click submit
before they're able to see their updated name. This is so that it's not
so jarring if they misspell their name.

Didn't make a lot of progress :/ I'm a lot rustier than I thought I was.
I may have to go back and do some serious review of Vue. In other words,
some major reVue.